backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
atpublic==4.1.0
attrs==23.2.0
bidict==0.23.1
bigframes==1.9.0
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
db-dtypes==1.2.0
decorator==5.1.1
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.6.0
gcsfs==2024.6.0
geopandas==1.0.0
google-api-core==2.19.1
google-auth==2.30.0
google-auth-oauthlib==1.2.0
google-cloud-aiplatform==1.56.0
google-cloud-bigquery==3.25.0
google-cloud-bigquery-connection==1.15.3
google-cloud-bigquery-storage==2.25.0
google-cloud-core==2.4.1
google-cloud-functions==1.16.3
google-cloud-iam==2.15.0
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
humanize==4.9.0
ibis-framework==8.0.0
idna==3.7
ipython==8.25.0
ipywidgets==8.1.3
jedi==0.19.1
jellyfish==1.0.4
Jinja2==3.1.4
joblib==1.4.2
jupyterlab_widgets==3.0.11
kiwisolver==1.4.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
multidict==6.0.5
multipledispatch==1.0.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
parsy==2.1
pexpect==4.9.0
pillow==10.3.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pydata-google-auth==1.8.2
Pygments==2.18.0
pyogrio==0.9.0
pyparsing==3.1.2
pyproj==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scikit-learn==1.5.0
scipy==1.14.0
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
sqlglot==20.11.0
stack-data==0.6.3
starlette==0.37.2
tabulate==0.9.0
termcolor==2.4.0
threadpoolctl==3.5.0
toolz==0.12.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
widgetsnbextension==4.0.11
yarl==1.9.4
